+++ sdk/resourcemanager/network/armnetwork/CHANGELOG.md
@@ -1,9 +1,11 @@
 # Release History
 
-## 4.3.0-beta.1 (2023-10-09)
+## 4.3.0 (2023-10-17)
 ### Features Added
 
-- Support for test fakes and OpenTelemetry trace spans.
+- New struct `InternetIngressPublicIPsProperties`
+- New field `InternetIngressPublicIPs` in struct `VirtualAppliancePropertiesFormat`
+
 
 ## 4.2.0 (2023-09-22)
 ### Features Added